refacto methods in concern
